requirements.txt
streamlit
pandas
plotly
python-dotenv
langchain-groq
langchain-core
openpyxl
langchain-openai
langchain-ollama
fastapi
uvicorn
jinja2
